pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rtl2gds_demo"
version = "0.0.1"
authors = [{ name = "Harry Wizard", email = "harry0789@qq.com" }]
description = "A tool to compile your RTL files into GDSII layouts."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
]
dependencies = ["pyyaml", "orjson", "klayout", "requests"]

[tool.pylint]
extension-pkg-whitelist = ["orjson"]

[project.urls]
Homepage = "https://atomgit.com/harrywh/rtl2gds"
Issues = "https://atomgit.com/harrywh/rtl2gds/issues"
Documentation = "https://atomgit.com/harrywh/rtl2gds"
Repository = "https://atomgit.com/harrywh/rtl2gds"

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
